Move multi-logger to own file
