+++ crates/re_log/src/lib.rs
@@ -23,6 +23,10 @@ pub use setup::*;
 
 pub use log::{Level, LevelFilter};
 
+mod multi_logger;
+
+pub use multi_logger::{add_boxed_logger, add_logger};
+
 #[cfg(target_arch = "wasm32")]
 mod log_web;
 
@@ -65,58 +69,10 @@ fn test_shorten_file_path() {
 
 // ----------------------------------------------------------------------------
 
-static MULTI_LOGGER: MultiLogger = MultiLogger::new();
-
-/// Install an additional global logger.
-pub fn add_boxed_logger(logger: Box<dyn log::Log>) {
-    add_logger(Box::leak(logger));
-}
-
-/// Install an additional global logger.
-pub fn add_logger(logger: &'static dyn log::Log) {
-    MULTI_LOGGER.loggers.write().push(logger);
-}
-
-/// Forward log messages to multiple [`log::log`] receivers.
-struct MultiLogger {
-    loggers: parking_lot::RwLock<Vec<&'static dyn log::Log>>,
-}
-
-impl MultiLogger {
-    pub const fn new() -> Self {
-        Self {
-            loggers: parking_lot::RwLock::new(vec![]),
-        }
-    }
-}
-
-impl log::Log for MultiLogger {
-    fn enabled(&self, metadata: &log::Metadata<'_>) -> bool {
-        self.loggers
-            .read()
-            .iter()
-            .any(|logger| logger.enabled(metadata))
-    }
-
-    fn log(&self, record: &log::Record<'_>) {
-        for logger in self.loggers.read().iter() {
-            logger.log(record);
-        }
-    }
-
-    fn flush(&self) {
-        for logger in self.loggers.read().iter() {
-            logger.flush();
-        }
-    }
-}
-
 pub struct LogMsg {
     pub level: Level,
     pub msg: String,
 }
-
-// ----------------------------------------------------------------------------
 
 /// Pipe log messages to a channel.
 pub struct ChannelLogger {
